player/earthwalker: fix compare signedness (3 fns to 100%)
- earthwalker_init: cast GameBit_Get result to int (signed cmpwi)
- fn_802A49C8: inline heldObj pointer checks (unsigned cmplwi, fresh reload)
- fn_802AAF80: u-suffix flag masks + u32 cast on bit test (cmplwi + rlwinm)

+++ src/main/dll/WC/dll_028A_wcearthwalker.c
@@ -509,7 +509,7 @@ void earthwalker_init(int obj, int setup)
     ewState->encounterType = *(u8*)(setup + 0x19);
     if (ewState->encounterType == 1)
     {
-        if (GameBit_Get(0x7fc) != 0 ||
+        if ((int)GameBit_Get(0x7fc) != 0 ||
             (*gMapEventInterface)->getMapAct(ewObj->mapEventId) == 2)
         {
             ewState->interactionState = 2;
